requirements.dev.txt
pytest-django==4.1.0
pytest-mock==3.5.1
python-dotenv==0.15.0